requirements.txt
onnxruntime==1.20.1
Pillow==11.1.0
transformers==4.48.3
